First version of class astrometry:
Initialized with a FITS HDU. Given the original WCS parameters from the header, projects stars from a reference catalog on the image, together with locations of stars detected in the image.

Also implemented routine filter_and_match, that allows matching and filtering the positions lists. Matching is done up to a given distance "tol" in degrees.
Should also implement and "exclusion" distance in degrees, that will ensure that the minimum distance between stars in the reference catalog is larger than "exclusion" in degrees